+++ spec/nova-common_spec.rb
@@ -250,6 +250,7 @@ describe 'openstack-compute::nova-common' do
         expect(chef_run).to render_file(file.name).with_content(/^key_file=$/)
         expect(chef_run).to render_file(file.name).with_content(/^cinder_ca_certificates_file=$/)
         expect(chef_run).to render_file(file.name).with_content(/^cinder_api_insecure=false/)
+        expect(chef_run).to render_file(file.name).with_content(/^cinder_catalog_info=volumev2:cinderv2:publicURL$/)
         expect(chef_run).to render_file(file.name).with_content(/^hash_algorithms = md5$/)
         expect(chef_run).to render_file(file.name).with_content(/^insecure = false$/)
         expect(chef_run).to render_file(file.name).with_content(/^glance_api_insecure=false$/)
@@ -286,8 +287,10 @@ describe 'openstack-compute::nova-common' do
       it 'sets cinder options' do
         node.set['openstack']['compute']['block-storage']['cinder_ca_certificates_file'] = 'dir/to/path'
         node.set['openstack']['compute']['block-storage']['cinder_api_insecure'] = true
+        node.set['openstack']['compute']['block-storage']['cinder_catalog_info'] = 'volume:cinder:publicURL'
         expect(chef_run).to render_file(file.name).with_content(/^cinder_api_insecure=true$/)
         expect(chef_run).to render_file(file.name).with_content(%r{^cinder_ca_certificates_file=dir/to/path$})
+        expect(chef_run).to render_file(file.name).with_content(/^cinder_catalog_info=volume:cinder:publicURL$/)
       end
 
       it 'sets memcached server(s)' do